qto-200302170620 update v0.8.1 docs, add dir handling in the cp-to-env shell action

+++ src/bash/qto/funcs/cp-to-env.func.sh
@@ -19,6 +19,9 @@ doCpToEnv(){
 
    relative_fpath=$(dirname "${file}")
    tgt_file=$tgt_PRODUCT_INSTANCE_DIR/$file
+   
    mkdir -p $(dirname $tgt_file)
-   cp -v $file $tgt_file
+   test -f $file && cp -v $file $tgt_file
+   test -d $tgt_file && rm -rv $tgt_file
+   test -d $file && rsync -v -X -r -E -o -g --perms --acls $file $(dirname $tgt_file)
 }
